% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{cross_named_lists}
\alias{cross_named_lists}
\title{Cross two named lists, preserving names}
\usage{
cross_named_lists(l, ...)
}
\arguments{
\item{l}{A named list of named lists}

\item{...}{Arguments to pass to \link[purrr]{cross_n}}
}
\value{
Returns a named list of named lists with all combinations of the
  first limits, their elements, named by combining the original list names
  with \code{sep}.
}
\description{
Cross two named lists, preserving names
}

