% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_artist_attributes}
\alias{dt_artist_attributes}
\title{Artist Attributes}
\format{A data frame with 37 rows and 5 variables:
\describe{
  \item{\code{artist}}{character. Artist name.}
  \item{\code{birth_early_year}}{integer. Earliest year of artist's birth.}
  \item{\code{birth_late_year}}{integer. Latest year of artist's birth.}
  \item{\code{death_early_year}}{integer. Earliest year of artist's death.}
  \item{\code{death_late_year}}{integer. Latest year of artist's death.}
}}
\usage{
dt_artist_attributes
}
\description{
Artist Attributes
}
\keyword{datasets}

