% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_compiled}
\alias{dt_compiled}
\title{Compiled model data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 588 rows and 138 columns.}
\usage{
dt_compiled
}
\description{
Compiled model data
}
\keyword{datasets}

