% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_motif_labels}
\alias{dt_motif_labels}
\title{Dutch tabletop motif labels}
\format{A data frame with 197 rows and 2 variables:
\describe{
  \item{\code{motif_code}}{character. A unique motif identifier.}
  \item{\code{motif_label}}{character. An English label for the motif code.}
}}
\usage{
dt_motif_labels
}
\description{
Dutch tabletop motif labels
}
\keyword{datasets}

