% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_motif_taxonomy}
\alias{dt_motif_taxonomy}
\title{Motif Taxonomy}
\format{A data frame with 394 rows and 2 variables:
\describe{
  \item{\code{motif_code}}{character. A unique motif identifier.}
  \item{\code{parent}}{character. The parent motif. For example, the motif code \code{Sh1a} (Oysters as a main course) is a child of \code{Sh1} (Oysters), and \code{Sh} (Shellfish).}
}}
\usage{
dt_motif_taxonomy
}
\description{
Motif Taxonomy
}
\keyword{datasets}

