% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_painting_artist}
\alias{dt_painting_artist}
\title{Artist-Painting relationships}
\format{A data frame with 120 rows and 3 variables:
\describe{
  \item{\code{painting_code}}{character. Join with \link{dt_paintings} \code{painting_code}.}
  \item{\code{artist}}{character. Join with \link{dt_artist_attributes} \code{artist}.}
  \item{\code{artist_relationship}}{character. The type of relationship: "Definite", "After", "Possible", "Co-painter".}
}}
\usage{
dt_painting_artist
}
\description{
Artist-Painting relationships
}
\keyword{datasets}

