% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_painting_motifs}
\alias{dt_painting_motifs}
\title{DATASET TITLE}
\format{A data frame with 1572 rows and 2 variables:
\describe{
  \item{\code{painting_code}}{character}
  \item{\code{motif_code}}{character}
}}
\usage{
dt_painting_motifs
}
\description{
DATASET DETAILS
}
\keyword{datasets}

