% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multiplot}
\alias{multiplot}
\title{Render ggplot while also writing multiple versions to disk}
\usage{
multiplot(p, filetypes = c("pdf", "svg", "png"), postfix = NULL,
  filenames = paste0("inst/plots/", filetypes, "/", deparse(substitute(p)),
  ifelse(is.null(postfix), "", paste0("_", postfix)), ".", filetypes), ...)
}
\arguments{
\item{p}{Plot object}

\item{filetypes}{Types of files to save}

\item{...}{Options to pass to \code{\link[ggplot2]{ggsave}}}

\item{filename}{Filenames}
}
\description{
Render ggplot while also writing multiple versions to disk
}

