% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation.R
\name{pca_rf}
\alias{pca_rf}
\title{Create a PCA of local importance measures from a random forest}
\usage{
pca_rf(rf, .filter = function(x) return(x))
}
\description{
Create a PCA of local importance measures from a random forest
}

