% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{prep_vars}
\alias{prep_vars}
\title{Prepare data frames for model building and prediction}
\usage{
prep_vars(x)
}
\arguments{
\item{x}{Data frame}

\item{rownames}{Quoted column name containting rownames. If NULL, no rownames are added.}
}
\description{
Cast strings and logical vectors as factors, and move specified column name to rownames
}

