% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{restore_rownames}
\alias{restore_rownames}
\title{Restore rownames}
\usage{
restore_rownames(d, col_name)
}
\arguments{
\item{d}{Data frame}

\item{col_name}{Quoted column name to be removed and written as rownames}
}
\description{
Convenince function to restore rownames to a data.frame from a specified
column. Rownames are usually eliminated by dplyr functions, but they have
their uses when computing models.
}
\seealso{
This is an inverse of \link[dplyr]{add_rownames}
}

