% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation.R
\name{rf_contested_obs}
\alias{rf_contested_obs}
\title{Identify the most contested observations in random forest model}
\usage{
rf_contested_obs(rf)
}
\arguments{
\item{rf}{randomForest object}
}
\value{
A dataframe with observation ID and vote diversity, ranked in descending order of diversity (e.g. with most contested observations at top)
}
\description{
In which observations do the trees in a mdoel most disagree about how to classfy?
}

