% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation.R
\name{rf_effects}
\alias{rf_effects}
\title{Plot effects of certain variables on the predicitons of a given random forest model}
\usage{
rf_effects(rf, class, terms = top_n_importance_names(rf = rf, class = class))
}
\arguments{
\item{rf}{A RandomForest object}

\item{class}{Character. The predicted class name}

\item{terms}{Character. Quoted term names. Defaults to the top 9 most important terms for the given variable as calculated by \link{top_n_importance_names}.}
}
\description{
Plot effects of certain variables on the predicitons of a given random forest model
}

