% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{run_rf}
\alias{run_rf}
\title{Run a randomForest model with helpful defaults for this particular dataset}
\usage{
run_rf(data, response, predictors, rownames = "painting_code", portion, tests,
  ntree = 2000, proximity = TRUE, localImp = TRUE, ...)
}
\arguments{
\item{data}{Dataframe to model.}

\item{response}{Quoted name of the response variable column.}

\item{predictors}{A character vector of predictor variables. This will be
combined with \code{response} in the format \code{as.function(response ~
predictor[1] + predictor[2] + ...)}. The value of \code{response} will be
removed from \code{predictors} if present.}

\item{rownames}{Quoted name of the column containing names of observations.}

\item{ntree}{Number of trees to build. Defaults to 2000.}

\item{proximity}{Whether to save proximity data or not. Defaults to TRUE.}

\item{localImp}{Whether to save local importance data. Defaults to TRUE.}

\item{...}{Other arguments to pass to
\code{\link[randomForest]{randomForest}}}
}
\value{
An object of class \code{randomForest}. See
  \code{\link[randomForest]{randomForest}}.
}
\description{
This convenience wrapper allows easy specification of different response and
predictor variables.
}

