% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_error}
\alias{theme_error}
\title{Error plot ggplot defaults}
\usage{
theme_error(xname, yname, padding = 0.1, high = 0.9, low = 0.1)
}
\arguments{
\item{xname}{The name of the predictor type on the x-axis}

\item{yname}{The name of the predictor type on the y-axis}

\item{padding}{Margin of quadrant labels.}
}
\description{
Adds a number of layers to a ggplot base to create a suitable dual-error comparison plot.
}

