% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation.R
\name{top_n_importance}
\alias{top_n_importance}
\alias{top_n_importance_names}
\title{The top n most important terms for a class}
\usage{
top_n_importance(rf, class, n)

top_n_importance_names(rf, class, n)
}
\arguments{
\item{rf}{RandomForest object}

\item{class}{Name of the category}

\item{n}{Integer. Number of topmost terms to return}
}
\value{
A data frame with term names and importance values
}
\description{
The top n most important terms for a class
}
\section{Functions}{
\itemize{
\item \code{top_n_importance_names}: Return the names as a vector only
}}

